fix(progressbar): move hasClass function in the helper-utils.spec in test utils folder
