+++ projects/igniteui-angular/src/lib/progressbar/circularbar.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IgxCircularProgressBarComponent } from './progressbar.component';
 import { configureTestSuite } from '../test-utils/configure-suite';
-import { hasClass } from "./helper-utils.spec";
+import { hasClass } from "../test-utils/helper-utils.spec";
 
 describe('IgxCircularProgressBarComponent', () => {
     let fixture: ComponentFixture<IgxCircularProgressBarComponent>;
