--- projects/igniteui-angular/src/lib/progressbar/helper-utils.spec.ts
@@ -1,3 +1,0 @@
-export function hasClass(element: HTMLElement, className: string, expected: boolean) {
-    expect(element.classList.contains(className)).toBe(expected);
-}
