+++ projects/igniteui-angular/src/lib/progressbar/linearbar.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { IgxLinearProgressBarComponent } from './progressbar.component';
 import { configureTestSuite } from '../test-utils/configure-suite';
-import { hasClass } from "./helper-utils.spec";
+import { hasClass } from "../test-utils/helper-utils.spec";
 
 describe('IgxLinearProgressBarComponent', () => {
     let fixture: ComponentFixture<IgxLinearProgressBarComponent>;
